+++ blockfrost-client/src/Blockfrost/Client/Cardano/Network.hs
@@ -9,12 +9,12 @@ import Blockfrost.Client.Types
 import Blockfrost.Types
 
 
-networkClient :: Project -> NetworkAPI (AsClientT BlockfrostClient)
+networkClient :: MonadBlockfrost m => Project -> NetworkAPI (AsClientT m)
 networkClient = fromServant . _network . cardanoClient
 
-getNetworkInfo_ :: Project -> BlockfrostClient Network
+getNetworkInfo_ :: MonadBlockfrost m => Project -> m Network
 getNetworkInfo_ = _networkInfo . networkClient
 
 -- | Get detailed network information.
-getNetworkInfo:: BlockfrostClient Network
+getNetworkInfo :: MonadBlockfrost m => m Network
 getNetworkInfo = go getNetworkInfo_