+++ blockfrost-client/src/Blockfrost/Client/Types.hs
@@ -1,5 +1,7 @@
 -- | Client Types
 {-# LANGUAGE CPP #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 
 module Blockfrost.Client.Types
   ( BlockfrostClient
@@ -20,6 +22,10 @@ module Blockfrost.Client.Types
   , fromServant
   , tryError
   , def
+  , BlockfrostClientT (..)
+  , MonadBlockfrost (..)
+  , runBlockfrostClientT
+  , newClientConfig
   ) where
 
 import Control.Monad.Except
@@ -34,36 +40,80 @@ import Blockfrost.API
 import Blockfrost.Client.Core
 
 type ClientConfig = (ClientEnv, Project)
-type BlockfrostClient =
-  ExceptT BlockfrostError
-    (ReaderT ClientConfig IO)
 
-apiClient :: BlockfrostAPI (AsClientT BlockfrostClient)
-apiClient = genericClientHoist nt
-  where nt :: ClientM a -> BlockfrostClient a
-        nt act = do
-          (env, _proj) <- ask
-          liftIO (runClientM act env)
-            >>= either (throwError . fromServantClientError) pure
+newtype BlockfrostClientT m a = BlockfrostClientT {
+  unBlockfrostClientT
+    :: ExceptT BlockfrostError
+        (ReaderT ClientConfig m) a
+  } deriving
+      ( Functor
+      , Applicative
+      , Monad
+      , MonadIO
+      , MonadReader ClientConfig
+      , MonadError BlockfrostError
+      )
 
-api0Client :: Project -> BlockfrostV0API (AsClientT BlockfrostClient)
+type BlockfrostClient = BlockfrostClientT IO
+
+class MonadIO m => MonadBlockfrost m where
+  liftBlockfrostClient :: ClientM a -> m a
+  getConf :: m ClientConfig
+
+instance MonadIO m => MonadBlockfrost (BlockfrostClientT m) where
+  liftBlockfrostClient act = BlockfrostClientT $ do
+    (env, _proj) <- ask
+    liftIO (runClientM act env)
+      >>= either (throwError . fromServantClientError) pure
+  getConf = BlockfrostClientT ask
+
+instance MonadBlockfrost ClientM where
+  liftBlockfrostClient = id
+  getConf = newClientConfig
+
+instance MonadBlockfrost IO where
+  liftBlockfrostClient act = getConf >>= \(env, _prj) -> runClientM act env >>= either (error . show) pure
+  getConf = newClientConfig
+
+apiClient :: MonadBlockfrost m => BlockfrostAPI (AsClientT m)
+apiClient = genericClientHoist liftBlockfrostClient
+
+api0Client :: MonadBlockfrost m => Project -> BlockfrostV0API (AsClientT m)
 api0Client = fromServant . _apiV0 apiClient
 
 -- ** Client runner
 
--- | Run @BlockfrostClient@ monad, using provided @Project@
-runBlockfrost :: Project
-  -> BlockfrostClient a
+-- | Run @BlockfrostClientT@ monad in @IO@, using provided @Project@
+runBlockfrost
+  :: Project
+  -> BlockfrostClientT IO a
   -> IO (Either BlockfrostError a)
-runBlockfrost proj act = do
-  env <- newEnvByProject proj
+runBlockfrost = runBlockfrostClientT
+
+-- | Run @BlockfrostClientT@, using provided @Project@
+runBlockfrostClientT
+  :: MonadIO m => Project
+  -> BlockfrostClientT m a
+  -> m (Either BlockfrostError a)
+runBlockfrostClientT proj act = do
+  env <- liftIO $ newEnvByProject proj
   flip runReaderT (env, proj)
-    $ runExceptT act
+    $ runExceptT $ unBlockfrostClientT act
+
+-- | Build default `ClientConfig` using BLOCKFROST_TOKEN_PATH environment variable
+newClientConfig
+  :: MonadIO m
+  => m ClientConfig
+newClientConfig = liftIO $ do
+  prj <- projectFromEnv
+  env <- newEnvByProject prj
+  pure (env, prj)
 
 -- | Helper
-go :: (Project -> BlockfrostClient a)
-  -> BlockfrostClient a
-go act = ask >>= act  . snd
+go :: MonadBlockfrost m
+   => (Project -> m a)
+   -> m a
+go act = getConf >>= act . snd
 
 -- Until mtl > 2.2.2
 -- https://github.com/haskell/mtl/pull/66
@@ -75,14 +125,14 @@ tryError action = (Right <$> action) `catchError` (pure . Left)
 
 -- ** Service clients
 
-commonClient :: Project -> CommonAPI (AsClientT BlockfrostClient)
+commonClient :: MonadBlockfrost m => Project -> CommonAPI (AsClientT m)
 commonClient = fromServant . _common . api0Client
 
-cardanoClient :: Project -> CardanoAPI (AsClientT BlockfrostClient)
+cardanoClient :: MonadBlockfrost m => Project -> CardanoAPI (AsClientT m)
 cardanoClient = fromServant . _cardano . api0Client
 
-ipfsClient :: Project -> IPFSAPI (AsClientT BlockfrostClient)
+ipfsClient :: MonadBlockfrost m => Project -> IPFSAPI (AsClientT m)
 ipfsClient = fromServant . _ipfs . api0Client
 
-nutLinkClient :: Project -> NutLinkAPI (AsClientT BlockfrostClient)
+nutLinkClient :: MonadBlockfrost m => Project -> NutLinkAPI (AsClientT m)
 nutLinkClient = fromServant . _nutLink . api0Client